init commit

+++ src/main/java/com/example/server2/familyEntity/FamilyRepository.java
@@ -3,10 +3,8 @@ package com.example.server2.familyEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface FamilyRepository extends CrudRepository<Family, Integer> {
+public interface FamilyRepository extends CrudRepository<Family, Long> {
 
 }
